backend/app/api/v1/dashboard.py
from fastapi import APIRouter, Depends, HTTPException
from typing import Dict, Any
from app.services.cache import get_revenue_summary
from app.core.auth import authenticate_request as get_current_user

router = APIRouter()

@router.get("/dashboard/summary")
async def get_dashboard_summary(
    property_id: str,
    current_user: dict = Depends(get_current_user)
) -> Dict[str, Any]:
    
    tenant_id = getattr(current_user, "tenant_id", "default_tenant") or "default_tenant"
    
    from datetime import datetime
    from dateutil.relativedelta import relativedelta
    
    # Calculate start and end of current month using dateutil for robustness
    # This handles leap years and month rollovers automatically and is more readable
    now = datetime.now()
    start_date = now.replace(day=1, hour=0, minute=0, second=0, microsecond=0)
    end_date = start_date + relativedelta(months=1)
    
    # Pass dates to service
    revenue_data = await get_revenue_summary(property_id, tenant_id, start_date, end_date)
    
    total_revenue_float = float(revenue_data['total'])
    
    return {
        "property_id": revenue_data['property_id'],
        "total_revenue": total_revenue_float,
        "currency": revenue_data['currency'],
        "reservations_count": revenue_data['count']
    }
