backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "aioredis>=2.0.0",
    "beautifulsoup4>=4.13.4",
    "cachetools>=6.2.0",
    "cryptography>=46.0.2",
    "fastapi>=0.116.1",
    "lz4>=4.4.4",
    "openai>=2.1.0",
    "orjson>=3.11.1",
    "pandas>=2.3.2",
    "pillow==10.2.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pyhumps>=3.8.0",
    "python-dateutil>=2.9.0.post0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "qrcode>=8.2",
    "redis>=5.0.0",
    "reportlab>=4.4.3",
    "requests>=2.32.4",
    "sendgrid>=6.12.4",
    "sqlalchemy>=2.0.0",
    "stripe>=7.0.0",
    "supabase>=2.18.0",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.35.0",
    "wheel==0.45.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
